댓글구현

+++ src/main/java/com/sparta/board/service/CommentService.java
@@ -6,27 +6,23 @@ import com.sparta.board.dto.MsgResponseDto;
 import com.sparta.board.entity.Board;
 import com.sparta.board.entity.Comment;
 import com.sparta.board.entity.User;
-import com.sparta.board.jwt.JwtUtil;
+import com.sparta.board.entity.UserRoleEnum;
 import com.sparta.board.repository.BoardRepository;
 import com.sparta.board.repository.CommentRepository;
-import com.sparta.board.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
 public class CommentService {
-
-    private final JwtUtil jwtUtil;
-
-    private final UserRepository userRepository;
-
     private final CommentRepository commentRepository;
 
     private final BoardRepository boardRepository;
 
+    //작성
     @Transactional
     public CommentResponseDto createComment(Long id, CommentRequestDto requestDto, User user) {
             Board board = boardRepository.findById(id).orElseThrow(
@@ -39,25 +35,37 @@ public class CommentService {
             return new CommentResponseDto(comment);
 
     }
+    //수정
     @Transactional
-    public CommentResponseDto updateComment(Long id, CommentRequestDto requestDto, User user) {
-
+    public ResponseEntity<?> updateComment(Long id, CommentRequestDto requestDto, User user) {
         Comment comment = commentRepository.findById(id).orElseThrow(
                 ()-> new IllegalArgumentException("없는댓글입니다.")
         );
-
-        comment.update(requestDto);
-
-        return new CommentResponseDto(comment);
+        if(user.getRole() == UserRoleEnum.ADMIN){
+            comment.update(requestDto);
+        } else {
+            if(!comment.getUser().getUsername().equals(user.getUsername())){
+                return ResponseEntity.ok(new MsgResponseDto("작성자만 수정할 수 있습니다.", HttpStatus.BAD_REQUEST.value()));
+            }
+            comment.update(requestDto);
+        }
+        return ResponseEntity.ok(new CommentResponseDto(comment));
 
     }
-
+    //삭제
     @Transactional
     public MsgResponseDto deleteComment(Long id, User user) {
         Comment comment = commentRepository.findById(id).orElseThrow(
                 ()-> new IllegalArgumentException("없는댓글입니다.")
         );
-        commentRepository.deleteById(id);
+        if(user.getRole() == UserRoleEnum.ADMIN){
+            commentRepository.deleteById(id);
+        } else{
+            if(!comment.getUser().getUsername().equals(user.getUsername())){
+                return new MsgResponseDto("작성자만 삭제할 수 있습니다.", HttpStatus.BAD_REQUEST.value());
+            }
+            commentRepository.deleteById(id);
+        }
         return new MsgResponseDto("댓글삭제완료", HttpStatus.OK.value());
     }
 }
